Minor fix - additional refactoring and modularize
