+++ lib/components/compass/compass_stream.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_background_service/flutter_background_service.dart';
 import 'package:flutter_compass/flutter_compass.dart';
-import 'package:gps/components/gps/gps_direction.dart';
+import 'package:gps/utility/direction_to_text.dart';
 
 double compDegree = 0;
 String compText = 'null';
@@ -10,7 +10,7 @@ void onStartCompass(ServiceInstance service) {
   FlutterCompass.events?.listen((event) {
     double? compassHeading = event.heading;
     compDegree = (compassHeading! + 360) % 360;
-    compText = GpsDirectionModule().directionToText(compDegree);
+    compText = directionToText(compDegree);
     service.invoke(
       'update_compass',
       {
