+++ lib/components/log/log_frame.dart
